gestion_image

+++ controllers/stuff.js
@@ -1,26 +1,63 @@
 const Books = require('../models/books');
+const fs = require('fs');
 
-exports.createBooks = (req, res, next) => {
-    delete req.body._id;
-    const postbooks = new Books({
-        ...req.body
+exports.createBook = (req, res, next) => {
+    console.log(req.file)
+    const bookObject = JSON.parse(req.body.book); // converti du json en js
+    console.log(bookObject)
+    delete bookObject._id;
+    delete bookObject._userId;
+    const book = new Books({
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-    postbooks.save()
-        .then(() => res.status(201).json({ message: 'Objet enregistré !' }))
-        .catch(error => res.status(400).json({ error }));
-}
+
+    book.save()
+        .then(() => { res.status(201).json({ message: 'Objet enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
+};
 
 exports.modifyBook = (req, res, next) => {
-    Books.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch(error => res.status(400).json({ error }));
-}
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
 
-exports.deleteBooks = (req, res, next) => {
-    Books.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
-}
+    delete bookObject._userId;
+    Books.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                Books.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+};
+
+exports.deleteBook = (req, res, next) => {
+    Books.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Books.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
 
 exports.getOneBook = (req, res, next) => {
     Books.findOne({ _id: req.params.id })
@@ -32,4 +69,13 @@ exports.getAllBooks = (req, res, next) => {
     Books.find()
         .then(book => res.status(200).json(book))
         .catch(error => res.status(400).json({ error }))
+}
+
+exports.getThreeBooks = (req, res, next) => {
+    const BooksSort = Books.sort((a, b) => b.averageRating - a.averageRating)
+    console.log(BooksSort)
+    const BooksThree = BooksSort.slice(0, 2)
+    console.log(BooksThree)
+        .then(book => res.status(200).json(book))
+        .catch(error => res.status(408).json({ error }))
 }
